requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
groq>=0.9.0
python-dotenv>=1.0.0
playwright>=1.44.0
pytest>=8.0.0
pytest-playwright>=0.5.0
pytest-bdd>=7.0.0
jinja2>=3.1.0
pandas>=2.0.0